+++ example/lib/local_notification_example.dart
@@ -51,7 +51,7 @@ class ChildWidget extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Container(
-        child: FlatButton(
+        child: TextButton(
             child: Text("Generate error"), onPressed: () => generateError()));
   }
 
@@ -124,5 +124,5 @@ class NotificationReportMode extends ReportMode {
 
   @override
   List<PlatformType> getSupportedPlatforms() =>
-      [PlatformType.Android, PlatformType.iOS];
+      [PlatformType.android, PlatformType.iOS];
 }