+++ example/lib/sentry_example.dart
@@ -48,7 +48,7 @@ class ChildWidget extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Container(
-        child: FlatButton(
+        child: TextButton(
             child: Text("Generate error"), onPressed: () => generateError()));
   }