+++ lib/handlers/email_manual_handler.dart
@@ -61,7 +61,7 @@ class EmailManualHandler extends ReportHandler {
   }
 
   String _getTitle(Report report) {
-    if (emailTitle != null && emailTitle.length > 0) {
+    if (emailTitle?.isNotEmpty == true) {
       return emailTitle;
     } else {
       return "Error report: >> ${report.error} <<";
@@ -77,8 +77,8 @@ class EmailManualHandler extends ReportHandler {
   }
 
   String _setupHtmlMessageText(Report report) {
-    StringBuffer buffer = StringBuffer("");
-    if (emailHeader != null && emailHeader.length > 0) {
+    final StringBuffer buffer = StringBuffer();
+    if (emailHeader?.isNotEmpty == true) {
       buffer.write(emailHeader);
       buffer.write("<hr><br>");
     }
@@ -93,21 +93,21 @@ class EmailManualHandler extends ReportHandler {
     }
     if (enableDeviceParameters) {
       buffer.write("<h2>Device parameters:</h2>");
-      for (var entry in report.deviceParameters.entries) {
+      for (final entry in report.deviceParameters.entries) {
         buffer.write("<b>${entry.key}</b>: ${entry.value}<br>");
       }
       buffer.write("<hr><br>");
     }
     if (enableApplicationParameters) {
       buffer.write("<h2>Application parameters:</h2>");
-      for (var entry in report.applicationParameters.entries) {
+      for (final entry in report.applicationParameters.entries) {
         buffer.write("<b>${entry.key}</b>: ${entry.value}<br>");
       }
       buffer.write("<br><br>");
     }
     if (enableCustomParameters) {
       buffer.write("<h2>Custom parameters:</h2>");
-      for (var entry in report.customParameters.entries) {
+      for (final entry in report.customParameters.entries) {
         buffer.write("<b>${entry.key}</b>: ${entry.value}<br>");
       }
       buffer.write("<br><br>");
@@ -117,8 +117,8 @@ class EmailManualHandler extends ReportHandler {
   }
 
   String _setupRawMessageText(Report report) {
-    StringBuffer buffer = StringBuffer("");
-    if (emailHeader != null && emailHeader.length > 0) {
+    final StringBuffer buffer = StringBuffer();
+    if (emailHeader?.isNotEmpty == true) {
       buffer.write(emailHeader);
       buffer.write("\n\n");
     }
@@ -133,21 +133,21 @@ class EmailManualHandler extends ReportHandler {
     }
     if (enableDeviceParameters) {
       buffer.write("Device parameters:\n");
-      for (var entry in report.deviceParameters.entries) {
+      for (final entry in report.deviceParameters.entries) {
         buffer.write("${entry.key}: ${entry.value}\n");
       }
       buffer.write("\n\n");
     }
     if (enableApplicationParameters) {
       buffer.write("Application parameters:\n");
-      for (var entry in report.applicationParameters.entries) {
+      for (final entry in report.applicationParameters.entries) {
         buffer.write("${entry.key}: ${entry.value}\n");
       }
       buffer.write("\n\n");
     }
     if (enableCustomParameters) {
       buffer.write("Custom parameters:\n");
-      for (var entry in report.customParameters.entries) {
+      for (final entry in report.customParameters.entries) {
         buffer.write("${entry.key}: ${entry.value}\n");
       }
       buffer.write("\n\n");
@@ -164,5 +164,5 @@ class EmailManualHandler extends ReportHandler {
 
   @override
   List<PlatformType> getSupportedPlatforms() =>
-      [PlatformType.Android, PlatformType.iOS];
+      [PlatformType.android, PlatformType.iOS];
 }