+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.4.1
+* Added localization text in toast handler.
+* Fixed page report mode UI issues by Don Coleman (https://github.com/drcdev).
+* Added handleSilentError parameter in CatcherOptions.
+* [BREAKING_CHANGE] PlatformType values has been changed to lower camel case.
+* Lint fixes.
+
 ## 0.4.0
 * [BREAKING_CHANGE]: rootWidget parameter has been moved to optional parameter
 * Catcher requires now rootWidget or runAppFunction in setup